Explicit types, avoid float division
This shaves off a good amount of CPU cycles here and there.
Since the ESP32 FPU doesn't have division, this is done in software.

GCC doesn't simplify division by a constant into multiplication
without using flags, so multiplication is done explicitly in code
where possible.

Also specifying a floating point value implicitly makes it double
precision, which in many cases caused a sign extension, only for
it to be turned back into single precision before being returned.

+++ src/fuel_economy_meter/trip.h
@@ -11,14 +11,14 @@ class Trip
 public:
   void begin(void);
 
-  uint16_t getRpm(void);
+  uint_fast16_t getRpm(void);
   float getLiters(void);
   float getKm(void);
   float getKmh(void);
   float getEfficiency();
   float getLph(void);
 
-  volatile uint64_t latestInjectionTime = 0; /// The duty time of the last injection pulse
+  volatile uint32_t latestInjectionTime = 0; /// The duty time of the last injection pulse
 
 private:
   static Trip *sTrip;
